Move profile settings files to new package
Those files have classes that are used to edit the user's profile, not their account. I'll make a new activity that edits the user's account (email, password, logout, delete account).
