+++ Project/app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:exported="false" />
 
         <activity
-            android:name=".ui.settings.SettingsProfileActivity"
+            android:name=".ui.settings.profile.SettingsProfileActivity"
             android:exported="false"
             android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
             >
