+++ Project/app/src/main/res/layout/activity_settings_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.settings.SettingsProfileActivity"
+    tools:context=".ui.settings.profile.SettingsProfileActivity"
     >
 
     <androidx.appcompat.widget.Toolbar
